+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- Ledger resource
+- LedgerTransaction resource
 
 ## [0.13.0] - 2026-06-29
 ### Added